+++ js/modules/adminModule.js
@@ -14,7 +14,7 @@ import { KNOWN_ENTITY_TYPES } from '../core/exportImportService.js';
 import { EntityRepository } from '../core/entityRepository.js';
 import { connectorMetaRepository } from '../core/entities/connectorMetaRepository.js';
 import { acompanhaPlusConnector } from '../core/connectors/acompanhaPlusConnector.js';
-import { expansionConnector } from '../core/connectors/expansionConnector.js';
+import { expansionYouthConnector } from '../core/connectors/expansionYouthConnector.js';
 import { plumaConnector } from '../core/connectors/plumaConnector.js';
 import { corporateCollectorConnector } from '../core/connectors/corporateCollectorConnector.js';
 import { jobSourceConnector } from '../core/connectors/jobSourceConnector.js';
@@ -25,7 +25,7 @@ import { estimateStorage } from '../core/db.js';
 import { clearAllModulesData } from '../core/dataManagementService.js';
 import { listRoles, listAssignableRoleNames, createRole, deleteRole, getRolePermissionsMap, setRolePermission, ALL_MODULE_KEYS } from '../core/roleService.js';
 
-const CONNECTORS = [acompanhaPlusConnector, expansionConnector, plumaConnector, corporateCollectorConnector, jobSourceConnector, jiraConnector];
+const CONNECTORS = [acompanhaPlusConnector, expansionYouthConnector, plumaConnector, corporateCollectorConnector, jobSourceConnector, jiraConnector];
 
 export async function render(container, ctx) {
   const { user, subview } = ctx;